Move source.go into activator/steplib
activateStepSourceFromModel's only caller is activator/steplib, and it
reaches the stepman package only through exported helpers (ReadRoute,
GetStepCacheDirPath, and the download primitive). Relocating source.go
there puts source activation next to its caller and lets the function be
unexported.

Export stepman.DownloadStepArchive (was downloadStepArchive) so the moved
code can reach the shared zip/git fetch primitive; DownloadStep still
delegates to it, so V1 behavior is unchanged. Comments tightened.
